requirements.txt
selenium==4.3.0
tauk==2.1.3
webdriver-manager==3.8.1
python-dotenv==0.20.0